+++ README.md
@@ -35,6 +35,6 @@
 
 - [x] 트랜잭션 경계 설정하기
 - [x] 트랜잭션 동기화 적용하기
-- [ ] 트랜잭션 서비스 추상화하기
+- [x] 트랜잭션 서비스 추상화하기
 
 </details>
